requirements.txt
pandas
numpy
scipy
matplotlib
networkx
python-libsbml